+++ app/src/main/java/com/infinity/omos/ui/music/search/SearchState.kt
@@ -1,4 +1,4 @@
-package com.infinity.omos.ui.music
+package com.infinity.omos.ui.music.search
 
 enum class SearchState {
     BEFORE, ING, AFTER
